+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `KeyPair`
+  * `from_secret/1` for getting a key pair from a secret
+  * `random/0` for generating a new key pair
 
 ## [0.1.1] - 2018-01-27
 ### Fixed